Move RequestArguments types to top level
